requirements.txt
aiogram==3.22.0
fastapi==0.118.0
uvicorn==0.37.0
pdfplumber==0.11.7
python-docx==1.2.0
python-dotenv==1.0.1
